==> scripts/00_Scraping.ipynb <==
# Scraping articles 
source: https://femminicidioitalia.info/lista/2022


from selenium import webdriver 
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.by import By
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.chrome.service import Service
import time 
import os 
import wget 
import getpass
import pandas as pd

s= Service('\Applications\chromedriver')

How to set up selenium : https://www.youtube.com/watch?v=95OqqfGi2Eg

#Optional- set options
chromeOptions = Options()
chromeOptions.headless= False 

#initialize driver 

driver= webdriver.Chrome(service=s, options= chromeOptions)

def open_website(): 
    '''
    opens website and clicks on cookie banner
    '''
    driver.get("https://femminicidioitalia.info/")
    print("opened")
    #cookie banner 
    time.sleep(5)
    cookie1=driver.find_element(By.XPATH, "//button[@class='fc-button fc-cta-consent fc-primary-button']")
    cookie1.click()

open_website()
    

#try for [year] : this opens the right list
year=driver.find_element(By.XPATH, "//a[@title='Lista dei femminicidi in Italia nel 2021']")
year.click()
#find all victims 
victims=driver.find_elements(By.CLASS_NAME, "victimName")
#check if it worked 
print(type(victims))
nvic=len(victims)
print(f"in 2021 there are articles about {nvic} victims.")

#create dataframe 
df1= pd.DataFrame(columns=('victim', 'multiple', 'killer', 'presumed', 'location', 'date', 'links' ))
df2= pd.DataFrame(columns=('victim', 'multiple', 'killer', 'presumed', 'location', 'date', 'links' ))


df1= pd.DataFrame(columns=('victim', 'multiple' , 'links' ))
df2= pd.DataFrame(columns=('victim', 'multiple',  'links' ))


#test for 1 
test=victims[25]
test.click()
time.sleep(5)
el=driver.find_elements(By.XPATH, "//span[@class='recapvalue fontspecialelite']")

driver.back()
#works 

def victim_info(): 
    '''
    on the page of the victim takes information and links
    '''
    ti=driver.find_elements(By.XPATH, "//span[@class='recaplabel fontspecialelite alignright']")
    el=driver.find_elements(By.XPATH, "//span[@class='recapvalue fontspecialelite']")

    for t, e in zip(ti, el):
        #print(t.text, e.text)
        if t.text=="Vittima:":
            df1['victim']=[e.text]
            df1['multiple']=0
        elif t.text=="Vittime:":
            df1['victim']=[e.text]
            df1['multiple']=1
       # elif t.text=="Presunto killer:":
            #df1['killer']=[e.text]
            #df1['presumed']=1
       # elif t.text=="Killer:":
           # df1['killer']=[e.text]
          #  df1['presumed']=0
       # elif t.text=="Località:":
          #  df1['location']=[e.text]
        #elif t.text=="Provincia": 
           # df1['province']=[e.text]
       # elif t.text=="Decesso:":
           # df1['death_loc']=[e.text]
        #elif t.text=="Data:":
         #   df1['date']=[e.text]
       # else: 
           # print("Missing info:", e.text)
    #print(df1)

                  

cols = ['victim', 'multiple', 'killer', 'presumed', 'location', 'death_loc', 'province', 'date', 'links' ]
lst = []

def victim_info2(): 
    '''
    on the page of the victim takes information and links
    '''
    ti=driver.find_elements(By.XPATH, "//span[@class='recaplabel fontspecialelite alignright']")
    el=driver.find_elements(By.XPATH, "//span[@class='recapvalue fontspecialelite']")

    for t, e in zip(ti, el):
        lst.append([
        if t.text=="Vittima:":
            
        ])
        
for a in range(2):
    lst.append([1, 2, 3])
df1 = pd.DataFrame(lst, columns=cols)
df1


def newspaper_info():
    #get links
    hr=driver.find_elements(By.XPATH, "//a[@class='fem_sharesource fem_sourceMain']")
    links = [h.get_attribute('href') for h in hr]

    #get titles
    titles =[]
    for h in hr:
        titles.append(h.text)

    #get name of newspaper
    np=driver.find_elements(By.XPATH, "//cite")
    nps=[]
    for n in np:
        nps.append(n.text)
    
    
    info=[]
    for l, t, n in zip(links, titles, nps):
        li=[]
        li.append(l)
        li.append(t)
        li.append(n)
        info.append(li)
        #print (info)
    df1['links']=[info]
    

len(victims)


count=0
for v in victims:
    count+=1
    print(count)
    try: 
        v.click()
    except: 
        try:
            v.click()
        except: 
            print(count)
    time.sleep(7)
    victim_info()
    time.sleep(5)
    newspaper_info()
    df2=pd.concat([df1, df2], axis=0, ignore_index=True)
    #print(df1)
    driver.back()
    time.sleep(7)


df2


df21_1


df2.to_csv("df2021.csv", index=False)
