==> scripts/02_Summarystatistics.Rmd <==
---
title: "Summary statistics"
author: "Francesca Giacco"
date: "2023-03-26"
output: html_document
---
```{r}
#load packages 

library(tidyverse)
library(ggplot2)
library(forcats)
library(stm)
library(quanteda)
library(quanteda.textstats)
library(stringr)
```

```{r}
#Summary statistics of dataset with info 


#n of femicides per region

a<- data21 %>%count(Regione) %>% 
  mutate(Regione = fct_reorder(as.factor(Regione), n)) %>% 
  ggplot( aes(x=n, y=Regione))+ geom_bar(stat='identity', fill="darkorchid4") +
    geom_text(aes(label=n), hjust=-0.2, size=3)+
    theme_minimal() + labs(title="Number of femicides per region", x="Number of femicides" , y="Region")
  
ggsave("fem_per_reg.png" ,plot= a, device="png", width=5, height=5 )


#origin of victim
  
a<- data21 %>%count(Paese.di.origine.persona.uccisa) %>% 
mutate(origin = fct_reorder(as.factor(Paese.di.origine.persona.uccisa), n))%>% 
  ggplot( aes(x=n, y=origin))+ geom_bar(stat='identity', fill="darkorchid4") +
    geom_text(aes(label=n), hjust=-0.2, size=3)+
    theme_minimal() + labs(title="Ethnicity of the victim", x="Number of victims" , y="Ethnicity")

a
  
ggsave("ethn_victim.png" ,plot= a, device="png", width=5, height=5 )



#origin killer

a<- data21 %>%count(Paese.di.origine.presunt..colpevole) %>% 
  mutate(origin = fct_reorder(as.factor(Paese.di.origine.presunt..colpevole), n))%>% 
  ggplot( aes(x=n, y=origin))+ geom_bar(stat='identity', fill="darkorchid4") +
    geom_text(aes(label=n), hjust=-0.2, size=3)+
    theme_minimal() + labs(title="Ethnicity of the killer", x="Number of culprits" , y="Ethnicity")

a
  
ggsave("ethn_kill.png" ,plot= a, device="png", width=5, height=5 )



#relationship to victim

a<- data21 %>%count(Relazione.con.la.persona.uccisa..semplificata.) %>% 
mutate(rel= fct_reorder(as.factor(Relazione.con.la.persona.uccisa..semplificata.), n))%>% 
  ggplot( aes(x=n, y=rel))+ geom_bar(stat='identity', fill="darkorchid4") +
    geom_text(aes(label=n), hjust=-0.2, size=3)+
    theme_minimal() + labs(title="Relationship to the victim", x="Count" , y="Relationship")

a
  
ggsave("relationship.png" ,plot= a, device="png", width=5, height=5 )



#how was the victim killed

a<- data21 %>%count(Causa.di.morte) %>% 
mutate(cause= fct_reorder(as.factor(Causa.di.morte), n))%>% 
  ggplot( aes(x=n, y=cause))+ geom_bar(stat='identity', fill="darkorchid4") +
    geom_text(aes(label=n), hjust=-0.2, size=3)+
    theme_minimal() + labs(title="Cause of death", x="Count" , y="Cause")

a
  
ggsave("cause.png" ,plot= a, device="png", width=5, height=5 )




#justification

table(data21$Giustificazione..movente) 

  
ggsave("justification.png" ,plot= a, device="png", width=5, height=5 )

```

# Summary statistics of datset with newspapers
```{r}
data_complete<-read_csv("merged - merged.csv") %>% 
  subset((!is.na(text))) 




#average n of articles per victim
data_complete %>%  
  count(victim) %>% 
  mutate(n1= n-1) %>% 
  subset(n1>0) %>% 
  summarise(mean(n1))


# share of local newspapers
sum(data_complete$local)/186

#journalists 
j<-data_complete %>%  subset((!is.na(journalist))) %>% 
  select(journalist, gender.j) %>% 
  count(gender.j) %>%  
  mutate(gender= c("M", "W", "1 W, 1 M", "2 W", "2 M, 1 W" )) %>% 
  

83/186

table(j$gender.j)

ggplot( j, aes(x=gender, y=n))+ geom_bar(stat='identity', fill="darkorchid4") +
    geom_text(aes(label=n), hjust=-0.2, size=3)+
    theme_minimal() + labs(title="Gender of journalist", x="Count" , y="Gender")


data_complete %>% count(picture)

data_complete$content
v<- c("p1", "p2", "p3", "p4")


cont<- data_complete %>% 
  select(content) %>% 
  separate(content, v,  ";" ) %>% 

pivot_longer(
    cols = starts_with("p"),
    names_to = "cont",
    names_prefix = "wk",
    values_to = "content",
    values_drop_na = TRUE
  ) %>% 
  mutate(content=gsub('[[:punct:] ]+', ' ',content))%>% 
  count(content) %>%  
  subset(n>2)



ggplot( cont, aes(x=content, y=n))+ geom_bar(stat='identity', fill="darkorchid4") +
    geom_text(aes(label=n), hjust=-0.2, size=3)+
    theme_minimal() + labs(title="Gender of journalist", x="Count" , y="Gender")



#topic modeling

texts<- data_complete %>% 
  mutate(text=tolower(text)) %>% 
  mutate(text=gsub('[[:punct:] ]+', ' ',text))%>% 
  rowwise() 

dfm_prep<- function(x) { 
  x$text%>% 
  tokens(remove_punct = TRUE, remove_numbers = TRUE) %>%
  tokens_tolower() %>%
  tokens(remove_punc=TRUE) %>% 
  tokens_remove(stopwords("Italian"))}


dfmm<- dfm_prep(texts) %>%
  tokens_wordstem(language = quanteda_options("language_stemmer")) %>% 
  tokens_remove(c("s")) %>% 
  dfm() %>% 
  dfm_trim(min_termfreq = 5)

dfmm

freq<- dfmm %>% textstat_frequency() %>% head(20)
freq$feature<- factor(freq$feature, levels=freq$feature)

a<- ggplot(freq, aes(x=frequency, y=feature, fill=docfreq)) +
 geom_col() + ggtitle("Most frequent words in the articles")
a

ggsave("freq_w.png", a,  width=6, height=6)
```

```{r}
#try STM

processed <- textProcessor(texts$text, metadata = texts)

out <- prepDocuments(processed$documents, processed$vocab, processed$meta)

docs <- out$documents
vocab <- out$vocab
meta <-out$meta

PrevFit <- stm(documents = out$documents, vocab = out$vocab, 
               K = 5, prevalence =~ text ,
                      max.em.its =75 , data = out$meta, gamma.prior='L1',
                       init.type = "Spectral", 
               seed=TRUE)

save(PrevFit, file = "model.RData")
load("model.RData")

labelTopics(PrevFit)
labelTopics(PrevFit, c(1:20))

```


```{r}
data<- read_csv("dws_clean.csv") 

#get comparison by nationality
data1<- data %>% 
          group_by(victim_origin) %>% 
                summarise(mean(text_vb))
data1

table(data$victim_origin)

data1<- data %>% 
          group_by(killer_origin) %>% 
                summarise(mean(text_vb))
data1


#get comparison by relationship
data1<- data %>% 
          group_by(Relazione.con.la.persona.uccisa..semplificata.) %>% 
                summarise(mean(text_vb))

data1

#gender of the journalist
data1<- data %>% 
          group_by(gender.j) %>% 
                summarise(mean(text_vb))

data1

data1<- data %>% 
          group_by(Regione) %>% 
                summarise(mean(text_vb))

table(data$killer_origin)


data1<- data %>% 
          group_by(local) %>% 
                summarise(mean(text_vb))

data1
```

# Check data at sentence level

```{r}
sent<- read_csv("datawscores_sent.csv")


sent<-sent %>% 
  mutate(sentence_vb= as.numeric(str_extract(sentence_vb, "[0-9].[0-9]+"))) 

sent%>% 
  ggplot(aes(x=sentence_vb))+ geom_histogram()

test<- sent %>% 
  subset(sentence_vb>0.50) 

test$sentence
```
# check data at paragraph level 

```{r}
np<- read_csv("npwscores.csv")

np <- np %>% 
  mutate(para_vb= as.numeric(str_extract(para_vb, "[0-9].[0-9]+"))) %>% 
  subset(nchar(paragraph)>50)
  

np %>%  ggplot(aes(x=para_vb))+ geom_histogram()

test<- np %>%  subset(para_vb>0.7)

test$paragraph



#topic modeling

texts<- np %>% 
  mutate(paragraph=tolower(paragraph))%>% 
  mutate(text=gsub('[[:punct:] ]+', ' ',paragraph))%>% 
  rowwise()




dfmm<- dfm_prep(texts) %>%
  tokens(remove_punct = TRUE, remove_numbers = TRUE) %>%
  tokens_tolower() %>%
  tokens(remove_punc=TRUE) %>% 
  tokens_remove(stopwords("italian")) %>% 
  tokens_wordstem(language = quanteda_options("language_stemmer")) %>% 
  tokens_remove(c("s")) %>% 
  dfm() %>% 
  dfm_trim(min_termfreq = 5)

dfmm

freq<- dfmm %>% textstat_frequency() %>% head(20)
freq$feature<- factor(freq$feature, levels=freq$feature)

a<- ggplot(freq, aes(x=frequency, y=feature, fill=docfreq)) +
 geom_col() + ggtitle("Most frequent words in the articles")
a

ggsave("freq_w.png", a,  width=6, height=6)
```

```{r}
#
#> tokens from 1 document.

processed <- textProcessor(texts$text, metadata = texts)

out <- prepDocuments(processed$documents, processed$vocab, processed$meta)

docs <- out$documents
vocab <- out$vocab
meta <-out$meta

PrevFit <- stm(documents = out$documents, vocab = out$vocab, 
               K = 15, prevalence =~ text ,
                      max.em.its =75 , data = out$meta, gamma.prior='L1',
                       init.type = "Spectral", 
               seed=TRUE)

save(PrevFit, file = "model.RData")
load("model.RData")

labelTopics(PrevFit)
labelTopics(PrevFit, c(1:20))



```
