==> scripts/06_models.Rmd <==
---
title: "Models"
author: "Francesca Giacco"
date: "2023-03-26"
output: html_document
---
```{r}
#load packages 
library(tidyverse)
library(ggplot2)
library(dplyr)
library(mgcv)
library(glmnet)
library(stargazer)
library(scales)
library(MASS)
library(sandwich)
library(lme4)
library(webshot)

#devtools::install_github("dustinfife/flexplot")
# install the development version
#devtools::install_github("dustinfife/flexplot", ref="development")
```


```{r}
#load data  (texts)

#upload labeled dataset: labeltest

data<- read_csv("labeltest.csv") %>% 
  subset(!is.na(vb_manual)) %>% 
  mutate(cvanda= ifelse(is.na(cvanda), 0, cvanda)) %>% 
  mutate(vb_manual2= vb_manual)
  #add layer of vb
  


a<- is.na(data$vb_manual)
table(a)
table(data$cvanda)

```


```{r}

#create 3 classes, binary and normalized column
data<- data %>%
  mutate(vb_3_1 = ifelse(vb_manual>=5, 1,
  ifelse (vb_manual <1, 0, 0.5))) %>%
  mutate(vb_bin = ifelse(vb_manual>=5, 1, 0)) %>%
  mutate(vb_3_2 = ifelse(vb_manual>5, 2,
  ifelse (vb_manual <3, 0, 1))) %>%
  mutate(vb_norm= rescale(vb_manual))

#redifining variables 
data<- data %>% mutate(vict_age_=ifelse(Età.persona.uccisa<14, "kid", 
             ifelse(Età.persona.uccisa<30, "teen/young_adult", 
             ifelse(Età.persona.uccisa<60,"adult","old" )))) %>% 
  mutate(kill_age_= ifelse(Età.presunt..colpevole<30, "teen/young_adult", 
             ifelse(Età.presunt..colpevole<60,"adult", "old"))) 

#cleaning for final plots 
data<- data %>% 
  mutate(relation_=relazione) %>% 
  mutate(journalist_=gender.j) %>% 
  mutate(pic_vict=cvict) %>% 
  mutate(pic_kill=cass) %>% 
  mutate(pic_both=cvanda) 
  
  
colnames(data)
```

```{r}
#write models: glm for manual_VB

glm0=glm(vb_manual~ killer_origin  + victim_origin + relation_, data=data, family=poisson)

glm1=glm(vb_manual~ killer_origin  + victim_origin +  + relation_  + vict_age_ + kill_age_, data=data, family=poisson)
glm2 =glm(vb_manual~ killer_origin  + victim_origin + relation_  + vict_age_ + kill_age_ + journalist_+ local, data=data, family=poisson)
#glm3=glm(vb_manual~ killer_origin  + victim_origin + relation_ + vict_age_+ kill_age_ + journalist_ + local+ Stupro..violenze.prima, data=data, family=poisson)



lst <- list(glm0, glm1, glm2)




stargazer(lst,  type = "html", out = "results/article_poisson.html", title="Regression of victim", align = TRUE)


#clustering: create clusters
#on last model


data<-data %>% group_by(title) %>% 
  mutate(title_number = group_indices()) %>% 
  ungroup()
# cluster standard errors by "cluster" variable
vcov0 <- vcovCL(glm0, cluster = data$title_number)
vcov1 <- vcovCL(glm1, cluster = data$title_number)
vcov2 <- vcovCL(glm2, cluster = data$title_number)
#vcov3 <- vcovCL(glm3, cluster = data$title_number)
# compute robust standard errors
se0 <- sqrt(diag(vcov0))
se1 <- sqrt(diag(vcov1))
se2 <- sqrt(diag(vcov2))
#se3<- sqrt(diag(vcov3))



stargazer(glm0,glm1, glm2, type = "html", out = "results/article_poisson_clustered.html", title="Regression of victim blaming with clustered SE (sentence level)", align = TRUE, 
          se = list(se0, se1, se2))

webshot::install_phantomjs()
webshot("results/article_poisson_clustered.html", file = "results/article_poisson_clustered.png")

table(data$previous_complaint)

```

```{r}
#binomial 

data<- data%>%  
  mutate(vb_bin=as.factor(vb_bin))#, levels=c(0,1), ordered=T)

glmbin0=glm(vb_bin~ killer_origin  + victim_origin + relation_, data=data, family=binomial)

glmbin1=glm(vb_bin~ killer_origin  + victim_origin +  + relation_  + vict_age_ + kill_age_, data=data, family=binomial)
glmbin2 =glm(vb_bin~ killer_origin  + victim_origin + relation_  + vict_age_ + kill_age_ + journalist_+ local, data=data, family=binomial)
  

lst <- list(glmbin0, glmbin1, glmbin2)




stargazer(kst,  type = "html", out = "results/article_binom.html")

table(data$vb_bin)

# cluster standard errors by "cluster" variable
vcov0 <- vcovCL(glmbin0, cluster = data$title_number)
vcov1 <- vcovCL(glmbin1, cluster = data$title_number)
vcov2 <- vcovCL(glmbin2, cluster = data$title_number)
#vcov3 <- vcovCL(glm3, cluster = data$title_number)
# compute robust standard errors
sbin0 <- sqrt(diag(vcov0))
sbin1 <- sqrt(diag(vcov1))
sbin2 <- sqrt(diag(vcov2))
#se3<- sqrt(diag(vcov3))



stargazer(glmbin0,glmbin1, glmbin2, type = "html", out = "results/sentence_binary_clustered.html", title="Regression of victim blaming with clustered SE and binary VB", align = TRUE, 
          se = list(sbin0, sbin1, sbin2))

webshot::install_phantomjs()
webshot("results/sentence_binary_clustered.html", file = "results/sentence_binary_clustered.png")


```

```{r}
#model for 3 classes 

data<- data%>%  
  mutate(vb_3_1=factor(vb_3_1, labels=1:3))#, levels=c(0,1), ordered=T)

polr0=polr(vb_3_1~ killer_origin  + victim_origin + relation_, data=data)

polr1=polr(vb_3_1~ killer_origin  + victim_origin +  + relation_  + vict_age_ + kill_age_, data=data)
polr2 =polr(vb_3_1~ killer_origin  + victim_origin + relation_  + vict_age_ + kill_age_ + journalist_+ local, data=data)
  

lst <- list(polr0, polr1, polr2)



stargazer(lst,  type = "html", out = "results/article_3class.html")


# cluster standard errors by "cluster" variable
vcov0 <- vcovCL(polr0, cluster = data$title_number)
vcov1 <- vcovCL(polr1, cluster = data$title_number)
vcov2 <- vcovCL(polr2, cluster = data$title_number)
#vcov3 <- vcovCL(glm3, cluster = data$title_number)
# compute robust standard errors
scl0 <- sqrt(diag(vcov0))
scl1 <- sqrt(diag(vcov1))
scl2 <- sqrt(diag(vcov2))
#se3<- sqrt(diag(vcov3))



stargazer(polr0,polr1, polr2, type = "html", out = "results/sentence_3classes_clustered.html", title="Regression of victim blaming with clustered SE and 3 classes VB", align = TRUE, se = list(scl0, scl1, scl2))



webshot("results/sentence_3classes_clustered.html", file = "results/sentence_3classes_clustered.png")


#combine binary and 3 classes at sentence level 

stargazer(polr0,polr1, polr2, glmbin0,glmbin1, glmbin2, type = "html", out = "results/sentence_3classesandbin_clustered.html", title="Regression of victim blaming with clustered SE and 3 classes VB", align = TRUE, se = list(scl0, scl1, scl2, sbin0, sbin1, sbin2))

webshot("results/sentence_3classesandbin_clustered.html", file = "results/sentence_3classesandbin_clustered.png")

```



```{r}
#to cluster standard errors 
#create clusters 
colnames(data)
data1<-data %>% group_by(title) %>% 
  mutate(title_number = group_indices()) %>% 
  ungroup()



glmer0<- glmer(vb_manual~ killer_origin  + victim_origin + relation_ + vict_age_+ kill_age_ + journalist_ + local,  data=data, family=poisson())

summary(glmer0)
```

```{r}
#at the article level
#i need to sum score for every article 
article<-data %>% group_by(title) %>% mutate(sum_vb=sum(vb_manual)) %>% ungroup() %>% 
  mutate(sum_vb= ifelse(cvanda>0, (vb_manual + 3), vb_manual))

article<- article[, -c(66:75)]

article<- unique(article)





a<- ggplot(article, aes(x=sum_vb)) + geom_bar(alpha=.3) + theme_bw() + labs(title="Distribution of manual scores",x ="Scores ", y = "Density")+ scale_color_brewer(palette="Dark2") + theme_bw()

a
table(article$sum_vb)

#poisson 
g0=glm(sum_vb~ killer_origin  + victim_origin + relation_, data=article, family=poisson)
summary(glm0)
g1=glm(sum_vb~ killer_origin  + victim_origin + relation_  + vict_age_ + kill_age_, data=article, family=poisson)
g2 =glm(sum_vb~ killer_origin  + victim_origin + relation_  + vict_age_ + kill_age_ + journalist_+ local, data=article, family=poisson)
#glm3 ="sum_vb~ killer_origin  + victim_origin + relation_ + vict_age_+ kill_age_ + journalist_ + local+  pic_vict +pic_both +pic_kill"
lst <- list(g0, g1, g2)

#models<- lst %>% map(~glm(., article, family=poisson()) )


stargazer(lst,  type = "html", out = "results/sum_article_poisson.html", title="Regression of victim", align = TRUE)


#add clustering 



#glmer0<- glmer(sum_vb~ killer_origin  + victim_origin + relation_ + (1 |victim) ,  data=article, family=poisson())

data<-data %>% group_by(victim) %>% 
  mutate(vict_number = group_indices()) %>% 
  ungroup()
# cluster standard errors by "cluster" variable
vcov0 <- vcovCL(g0, cluster = article$vict_number)
vcov1 <- vcovCL(g1, cluster = article$vict_number)
vcov2 <- vcovCL(g2, cluster = article$vict_number)
#vcov3 <- vcovCL(glm3, cluster = data$title_number)
# compute robust standard errors
s0 <- sqrt(diag(vcov0))
s1 <- sqrt(diag(vcov1))
s2 <- sqrt(diag(vcov2))
#se3<- sqrt(diag(vcov3))



stargazer(glm0,  glm1, glm2, g0, g1, g2,  type = "html", out = "results/combined_poisson_clustered.html", title="Regression of victim blaming with clustered SE (sentence and article level)", align = TRUE, 
          se = list(se0, se1, se2, s0, s1, s2))

webshot("results/combined_poisson_clustered.html", file = "results/combined_poisson_clustered.png")

```


```{r}
#robustnes check  article level

table(article$sum_vb)

article<- article %>%
  mutate(vb_4 = ifelse(sum_vb<10, 0,
  ifelse (sum_vb <30, 1 , 
  ifelse(sum_vb< 60, 2, 30))))
  #mutate(vb_bin = ifelse(vb_manual>=5, 1, 0)) %>%
  #mutate(vb_3_2 = ifelse(vb_manual>5, 2,
 # ifelse (vb_manual <3, 0, 1))) %>%
 # mutate(vb_norm= rescale(vb_manual))

article<- article%>%  
  mutate(vb_4=factor(vb_4, labels=1:4))#, levels=c(0,1), ordered=T)

p0=polr(vb_4~ killer_origin  + victim_origin + relation_, data=article)

p1=polr(vb_4~ killer_origin  + victim_origin + relation_  + vict_age_ + kill_age_, data=article)
p2 =polr(vb_4~ killer_origin  + victim_origin + relation_  + vict_age_ + kill_age_ + journalist_+ local, data=article)

lst <- list(p0, p1, p2)

# cluster standard errors by "cluster" variable
vcov0 <- vcovCL(p0, cluster = article$vict_number)
vcov1 <- vcovCL(p1, cluster = article$vict_number)
vcov2 <- vcovCL(p2, cluster = article$vict_number)
#vcov3 <- vcovCL(glm3, cluster = data$title_number)
# compute robust standard errors
se0 <- sqrt(diag(vcov0))
se1 <- sqrt(diag(vcov1))
se2 <- sqrt(diag(vcov2))
#se3<- sqrt(diag(vcov3))

stargazer(lst,  type = "html", out = "results/sum_article_3class_clustered.html", title="Regression of victim blaming with clustered SE (article level and 4 classes)", align = TRUE,  se = list(se0, se1, se2))
```

```{r}
#distribution of significant coefficients 

d<- ggplot(data, aes(x=vict_age_)) + geom_bar() + theme_bw() + labs(title=" Distribution of sentences across age groups",x ="Age group ", y="Number of article sentences")
d
data2 <- data %>% 
  group_by(vict_age_) %>% 
  summarise(sum(vb_manual))

e<- ggplot(data2, aes(x=vict_age_, y=`mean(vb_manual)`, width=0.4))+ geom_bar(stat="identity", fill=colors()[527])+  labs(title="Victim blaming by victim age",x ="Age group ", y = "Average victim blaming")+ theme_bw()

e

e<- ggplot(data, aes(x=relation_)) + geom_bar() + theme_bw() + labs(title=" Distribution of sentences across relationship types",x ="Type of relationship ", y="Number of article sentences")


a<- is.na(data$vb_manual)
table(a)
colnames(data)

```



`
